Extract functions for ZTP integration
To start an agent image, the assisted service needs to know the Ironic
URL and kernel parameters. This change makes the corresponding functions
public, moves them to a new utils.go and detaches from ProvisioningInfo.
