+++ provisioning/utils.go
@@ -1,0 +1,71 @@
+package provisioning
+
+import (
+	"context"
+	"fmt"
+	"net"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	coreclientv1 "k8s.io/client-go/kubernetes/typed/core/v1"
+
+	metal3iov1alpha1 "github.com/openshift/cluster-baremetal-operator/api/v1alpha1"
+)
+
+func getPodHostIP(podClient coreclientv1.PodsGetter, targetNamespace string) (string, error) {
+	labelSelector := &metav1.LabelSelector{
+		MatchLabels: map[string]string{
+			"k8s-app":    metal3AppName,
+			cboLabelName: stateService,
+		}}
+
+	selector, err := metav1.LabelSelectorAsSelector(labelSelector)
+	if err != nil {
+		return "", err
+	}
+
+	listOptions := metav1.ListOptions{
+		LabelSelector: selector.String(),
+	}
+
+	podList, err := podClient.Pods(targetNamespace).List(context.Background(), listOptions)
+	if err != nil {
+		return "", err
+	}
+
+	var hostIP string
+	switch len(podList.Items) {
+	case 0:
+		// Ironic IP not available yet, just return an empty string
+	case 1:
+		hostIP = podList.Items[0].Status.HostIP
+	default:
+		// We expect only one pod with the above LabelSelector
+		err = fmt.Errorf("there should be only one pod listed for the given label")
+	}
+
+	return hostIP, err
+}
+
+func GetIronicIP(client kubernetes.Interface, targetNamespace string, config *metal3iov1alpha1.ProvisioningSpec) (string, error) {
+	if config.ProvisioningNetwork != metal3iov1alpha1.ProvisioningNetworkDisabled && !config.VirtualMediaViaExternalNetwork {
+		return config.ProvisioningIP, nil
+	} else {
+		return getPodHostIP(client.CoreV1(), targetNamespace)
+	}
+}
+
+func IpOptionForProvisioning(config *metal3iov1alpha1.ProvisioningSpec, networkStack NetworkStackType) string {
+	var optionValue string
+	ip := net.ParseIP(config.ProvisioningIP)
+	if config.ProvisioningNetwork == metal3iov1alpha1.ProvisioningNetworkDisabled || ip == nil {
+		// It ProvisioningNetworkDisabled or no valid IP to check, fallback to the external network
+		return networkStack.IpOption()
+	}
+	if ip.To4() != nil {
+		optionValue = "ip=dhcp"
+	} else {
+		optionValue = "ip=dhcp6"
+	}
+	return optionValue
+}
